requirements.txt
DateTime>=5.5
python-dotenv>=1.0.1
fastapi>=0.111.0
jose>=1.0.0
python-jose>=3.4.0
SQLAlchemy>=2.0.31
requests>=2.32.3
httpx>=0.27.0
beautifulsoup4>=4.12.2
pydantic>=2.7.4
passlib>=1.7.4
typing_extensions>=4.12.2
psycopg2-binary>=2.9.3
fastapi[all]>=0.68.0